Don't increment COUNTER when expiring the backup.
This one is a bit subtle. Let's say you have several days worth of
hourly backups (from an old version of this script that doesn't do
expiry). Then you update to this and it begins expiring your oldest
backups. Well it wouldn't keep a month worth of dailies, because each
hourly backup older than the most recent 24 increments the counter,
but then gets deleted. So we have to decrement the counter variable
inside the expiry function in order to counteract that effect. This
way it will actually keep a month's worth of dailies.